Vulkan: Refactor SharedGarbageList into templated class
This CL mostly involves non-functional changes to prepare for next CL.
No behavior change is expected.

This CL wraps the garbage list into its own templated class which
maintains std::queue and tracks number bytes in the queue etc.

This CL also renames SharedBufferSuballocationGarbageList to
BufferSuballocationGarbageList to reduce verbosity a bit.

This CL deleted GarbageAndQueueSerial and GarbageQueue since they are no
longer being used.

This renames vk::GarbageList to vk::GarbageObjects to reduce name
confusion with SharedGarbageList.

Bug: b/302739073
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/4919816
Commit-Queue: Charlie Lao <cclao@google.com>

+++ src/libANGLE/renderer/vulkan/ContextVk.cpp
@@ -3385,7 +3385,8 @@ void ContextVk::syncObjectPerfCounters(const angle::VulkanPerfCounters &commandQ
     // Return current drawFramebuffer's cache stats
     mPerfCounters.framebufferCacheSize = mShareGroupVk->getFramebufferCache().getSize();
 
-    mPerfCounters.pendingSubmissionGarbageObjects = mRenderer->getPendingSubmissionGarbageSize();
+    mPerfCounters.pendingSubmissionGarbageObjects =
+        static_cast<uint64_t>(mRenderer->getPendingSubmissionGarbageSize());
 }
 
 void ContextVk::updateOverlayOnPresent()
